Rename interprocedural_cfg to expr_visitor, and base_cf to stmt_visitor
