+++ .pre-commit-config.yaml
@@ -3,18 +3,9 @@
     hooks:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
-    -   id: autopep8-wrapper
     -   id: check-docstring-first
     -   id: debug-statements
     -   id: check-ast
     -   id: check-symlinks
-    -   id: name-tests-test
-        exclude: tests/util
     -   id: flake8
         args: ['--ignore=E501']
-        exclude: ^test_data/
--   repo: https://github.com/asottile/reorder_python_imports
-    sha: v0.3.5
-    hooks:
-    -   id: reorder-python-imports
-        language_version: python3.6
